+++ src/main/java/io/vepo/morphoboard/ticket/StatusResource.java
@@ -11,11 +11,20 @@ import jakarta.ws.rs.core.MediaType;
 @Path("/api/statuses")
 @Produces(MediaType.APPLICATION_JSON)
 public class StatusResource {
+    public static final record StatusResponse(String name) {
+    }
+
+    private static StatusResponse toResponse(Status status) {
+        return new StatusResponse(status.name);
+    }
+
     @Inject
     StatusRepository repository;
 
     @GET
-    public List<Status> listAll() {
-        return repository.listAll();
+    public List<StatusResponse> listAll() {
+        return repository.streamAll()
+                         .map(StatusResource::toResponse)
+                         .toList();
     }
 }